Can show error page

+++ 3.00-starting-project/src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -34,6 +34,10 @@ public class GradebookController {
 
     @GetMapping("/delete/student/{id}")
     public String deleteStudent(@PathVariable int id, Model m) {
+        if(!studentService.checkIfStudentIsNull(id)) {
+            return "error";
+        }
+
         studentService.deleteStudent(id);
 
         Iterable<CollegeStudent> collegeStudents = studentService.getGradebook();
